added a new sqa loop tool called upgradeloop.sh that will boot upgrade loop.  Overhauled the init.d sqa script to support papertrail logging + ease of use

+++ dev-tools/SQA/nodefiller.sh
@@ -5,25 +5,31 @@ fsize=48M;
 
 #/etc/init.d/wdogreporter.sh &
 
+starttime="$1"
+if [[ $starttime = "" || upgradenum = "" ]]; then
+	echo "oom nodefiller is not going to begin.  useage: nodefiller.sh <starttime>"
+	echo "failed to provide: starttime: $starttime"
+else
 
-echo -e "-----\nstarting a new session\n" 
-echo -e "   sleeping 10"
-sleep 90
-for i in {1..5}; do
-	echo -e "  fake$i created"
-	fallocate -l $fsize /var/log/fake$i
-	df -h 
-	echo "   sleeping $timing"
-	sleep $timing
-done
-echo -e "done filling var/log"
-while true; do
-	echo -e "   firing up node " 
-	node &
-	npid1=$!
-	sleep $timing
-	echo "   my node: $npid1" 
-	sleep $timing
-	kill -9 $npid1
-	sleep 10
-done
+	echo -e "-----\nstarting a new session\n" 
+	echo -e "   sleeping $starttime"
+	sleep $starttime
+	for i in {1..5}; do
+		echo -e "  fake$i created"
+		fallocate -l $fsize /var/log/fake$i
+		df -h 
+		echo "   sleeping $timing"
+		sleep $timing
+	done
+	echo -e "done filling var/log"
+	while true; do
+		echo -e "   firing up node " 
+		node &
+		npid1=$!
+		sleep $timing
+		echo "   my node: $npid1" 
+		sleep $timing
+		kill -9 $npid1
+		sleep 10
+	done
+fi
